Move params to shared module space
