+++ modules/text2vec-contextionary/neartext/near_text_extract.go
@@ -11,12 +11,12 @@
 
 package neartext
 
-import "github.com/weaviate/weaviate/entities/searchparams"
+import "github.com/weaviate/weaviate/usecases/modulecomponents/nearText"
 
 // ExtractNearText arguments, such as "concepts", "moveTo", "moveAwayFrom",
 // "limit", etc.
 func (g *GraphQLArgumentsProvider) extractNearTextFn(source map[string]interface{}) interface{} {
-	var args searchparams.NearTextParams
+	var args nearText.NearTextParams
 
 	// keywords is a required argument, so we don't need to check for its existing
 	keywords := source["concepts"].([]interface{})
@@ -79,12 +79,12 @@ func (g *GraphQLArgumentsProvider) extractNearTextFn(source map[string]interface
 	return &args
 }
 
-func (g *GraphQLArgumentsProvider) extractMovement(input interface{}) searchparams.ExploreMove {
+func (g *GraphQLArgumentsProvider) extractMovement(input interface{}) nearText.ExploreMove {
 	// the type is fixed through gql config, no need to catch incorrect type
 	// assumption, all fields are required so we don't need to check for their
 	// presence
 	moveToMap := input.(map[string]interface{})
-	res := searchparams.ExploreMove{}
+	res := nearText.ExploreMove{}
 	res.Force = float32(moveToMap["force"].(float64))
 
 	keywords, ok := moveToMap["concepts"].([]interface{})
@@ -97,7 +97,7 @@ func (g *GraphQLArgumentsProvider) extractMovement(input interface{}) searchpara
 
 	objects, ok := moveToMap["objects"].([]interface{})
 	if ok {
-		res.Objects = make([]searchparams.ObjectMove, len(objects))
+		res.Objects = make([]nearText.ObjectMove, len(objects))
 		for i, value := range objects {
 			v, ok := value.(map[string]interface{})
 			if ok {
