+++ modules/text2vec-contextionary/neartext/near_text_extract_test.go
@@ -15,7 +15,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/weaviate/weaviate/entities/searchparams"
+	"github.com/weaviate/weaviate/usecases/modulecomponents/nearText"
 )
 
 func Test_extractNearTextFn(t *testing.T) {
@@ -25,7 +25,7 @@ func Test_extractNearTextFn(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want *searchparams.NearTextParams
+		want *nearText.NearTextParams
 	}{
 		{
 			"Extract with concepts",
@@ -34,7 +34,7 @@ func Test_extractNearTextFn(t *testing.T) {
 					"concepts": []interface{}{"c1", "c2", "c3"},
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values: []string{"c1", "c2", "c3"},
 			},
 		},
@@ -48,7 +48,7 @@ func Test_extractNearTextFn(t *testing.T) {
 					"network":  true,
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values:       []string{"c1", "c2", "c3"},
 				Distance:     0.4,
 				WithDistance: true,
@@ -66,7 +66,7 @@ func Test_extractNearTextFn(t *testing.T) {
 					"network":   true,
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values:    []string{"c1", "c2", "c3"},
 				Certainty: 0.4,
 				Limit:     100,
@@ -91,17 +91,17 @@ func Test_extractNearTextFn(t *testing.T) {
 					},
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values:       []string{"c1", "c2", "c3"},
 				Distance:     0.89,
 				WithDistance: true,
 				Limit:        500,
 				Network:      false,
-				MoveTo: searchparams.ExploreMove{
+				MoveTo: nearText.ExploreMove{
 					Values: []string{"positive"},
 					Force:  0.5,
 				},
-				MoveAwayFrom: searchparams.ExploreMove{
+				MoveAwayFrom: nearText.ExploreMove{
 					Values: []string{"epic"},
 					Force:  0.25,
 				},
@@ -125,16 +125,16 @@ func Test_extractNearTextFn(t *testing.T) {
 					},
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values:    []string{"c1", "c2", "c3"},
 				Certainty: 0.89,
 				Limit:     500,
 				Network:   false,
-				MoveTo: searchparams.ExploreMove{
+				MoveTo: nearText.ExploreMove{
 					Values: []string{"positive"},
 					Force:  0.5,
 				},
-				MoveAwayFrom: searchparams.ExploreMove{
+				MoveAwayFrom: nearText.ExploreMove{
 					Values: []string{"epic"},
 					Force:  0.25,
 				},
@@ -183,25 +183,25 @@ func Test_extractNearTextFn(t *testing.T) {
 					},
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values:       []string{"c1", "c2", "c3"},
 				Distance:     0.89,
 				WithDistance: true,
 				Limit:        500,
 				Network:      false,
-				MoveTo: searchparams.ExploreMove{
+				MoveTo: nearText.ExploreMove{
 					Values: []string{"positive"},
 					Force:  0.5,
-					Objects: []searchparams.ObjectMove{
+					Objects: []nearText.ObjectMove{
 						{ID: "moveTo-uuid1"},
 						{Beacon: "weaviate://localhost/moveTo-uuid2"},
 						{Beacon: "weaviate://localhost/moveTo-uuid3"},
 					},
 				},
-				MoveAwayFrom: searchparams.ExploreMove{
+				MoveAwayFrom: nearText.ExploreMove{
 					Values: []string{"epic"},
 					Force:  0.25,
-					Objects: []searchparams.ObjectMove{
+					Objects: []nearText.ObjectMove{
 						{ID: "moveAwayFrom-uuid1"},
 						{ID: "moveAwayFrom-uuid2"},
 						{Beacon: "weaviate://localhost/moveAwayFrom-uuid3"},
@@ -253,24 +253,24 @@ func Test_extractNearTextFn(t *testing.T) {
 					},
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values:    []string{"c1", "c2", "c3"},
 				Certainty: 0.89,
 				Limit:     500,
 				Network:   false,
-				MoveTo: searchparams.ExploreMove{
+				MoveTo: nearText.ExploreMove{
 					Values: []string{"positive"},
 					Force:  0.5,
-					Objects: []searchparams.ObjectMove{
+					Objects: []nearText.ObjectMove{
 						{ID: "moveTo-uuid1"},
 						{Beacon: "weaviate://localhost/moveTo-uuid2"},
 						{Beacon: "weaviate://localhost/moveTo-uuid3"},
 					},
 				},
-				MoveAwayFrom: searchparams.ExploreMove{
+				MoveAwayFrom: nearText.ExploreMove{
 					Values: []string{"epic"},
 					Force:  0.25,
-					Objects: []searchparams.ObjectMove{
+					Objects: []nearText.ObjectMove{
 						{ID: "moveAwayFrom-uuid1"},
 						{ID: "moveAwayFrom-uuid2"},
 						{Beacon: "weaviate://localhost/moveAwayFrom-uuid3"},
@@ -323,24 +323,24 @@ func Test_extractNearTextFn(t *testing.T) {
 					},
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values:       []string{"c1", "c2", "c3"},
 				Distance:     0.89,
 				WithDistance: true,
 				Limit:        500,
 				Network:      false,
-				MoveTo: searchparams.ExploreMove{
+				MoveTo: nearText.ExploreMove{
 					Values: []string{"positive"},
 					Force:  0.5,
-					Objects: []searchparams.ObjectMove{
+					Objects: []nearText.ObjectMove{
 						{ID: "moveTo-uuid1", Beacon: "weaviate://localhost/moveTo-uuid2"},
 						{ID: "moveTo-uuid1", Beacon: "weaviate://localhost/moveTo-uuid2"},
 					},
 				},
-				MoveAwayFrom: searchparams.ExploreMove{
+				MoveAwayFrom: nearText.ExploreMove{
 					Values: []string{"epic"},
 					Force:  0.25,
-					Objects: []searchparams.ObjectMove{
+					Objects: []nearText.ObjectMove{
 						{ID: "moveAwayFrom-uuid1", Beacon: "weaviate://localhost/moveAwayFrom-uuid1"},
 						{ID: "moveAwayFrom-uuid2", Beacon: "weaviate://localhost/moveAwayFrom-uuid2"},
 						{Beacon: "weaviate://localhost/moveAwayFrom-uuid3"},
@@ -393,23 +393,23 @@ func Test_extractNearTextFn(t *testing.T) {
 					},
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values:    []string{"c1", "c2", "c3"},
 				Certainty: 0.89,
 				Limit:     500,
 				Network:   false,
-				MoveTo: searchparams.ExploreMove{
+				MoveTo: nearText.ExploreMove{
 					Values: []string{"positive"},
 					Force:  0.5,
-					Objects: []searchparams.ObjectMove{
+					Objects: []nearText.ObjectMove{
 						{ID: "moveTo-uuid1", Beacon: "weaviate://localhost/moveTo-uuid2"},
 						{ID: "moveTo-uuid1", Beacon: "weaviate://localhost/moveTo-uuid2"},
 					},
 				},
-				MoveAwayFrom: searchparams.ExploreMove{
+				MoveAwayFrom: nearText.ExploreMove{
 					Values: []string{"epic"},
 					Force:  0.25,
-					Objects: []searchparams.ObjectMove{
+					Objects: []nearText.ObjectMove{
 						{ID: "moveAwayFrom-uuid1", Beacon: "weaviate://localhost/moveAwayFrom-uuid1"},
 						{ID: "moveAwayFrom-uuid2", Beacon: "weaviate://localhost/moveAwayFrom-uuid2"},
 						{Beacon: "weaviate://localhost/moveAwayFrom-uuid3"},
@@ -423,7 +423,7 @@ func Test_extractNearTextFn(t *testing.T) {
 	testsWithAutocorrect := []struct {
 		name string
 		args args
-		want *searchparams.NearTextParams
+		want *nearText.NearTextParams
 	}{
 		{
 			"Extract with concepts",
@@ -433,7 +433,7 @@ func Test_extractNearTextFn(t *testing.T) {
 					"autocorrect": true,
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values:      []string{"c1", "c2", "c3"},
 				Autocorrect: true,
 			},
@@ -446,7 +446,7 @@ func Test_extractNearTextFn(t *testing.T) {
 					"autocorrect": true,
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values:      []string{"transformed text", "c2", "transformed text"},
 				Autocorrect: true,
 			},
@@ -496,25 +496,25 @@ func Test_extractNearTextFn(t *testing.T) {
 					},
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values:       []string{"transformed text", "c1", "c2", "c3", "transformed text"},
 				Distance:     0.89,
 				WithDistance: true,
 				Limit:        500,
 				Network:      false,
 				Autocorrect:  true,
-				MoveTo: searchparams.ExploreMove{
+				MoveTo: nearText.ExploreMove{
 					Values: []string{"positive"},
 					Force:  0.5,
-					Objects: []searchparams.ObjectMove{
+					Objects: []nearText.ObjectMove{
 						{ID: "moveTo-uuid1", Beacon: "weaviate://localhost/moveTo-uuid2"},
 						{ID: "moveTo-uuid1", Beacon: "weaviate://localhost/moveTo-uuid2"},
 					},
 				},
-				MoveAwayFrom: searchparams.ExploreMove{
+				MoveAwayFrom: nearText.ExploreMove{
 					Values: []string{"epic"},
 					Force:  0.25,
-					Objects: []searchparams.ObjectMove{
+					Objects: []nearText.ObjectMove{
 						{ID: "moveAwayFrom-uuid1", Beacon: "weaviate://localhost/moveAwayFrom-uuid1"},
 						{ID: "moveAwayFrom-uuid2", Beacon: "weaviate://localhost/moveAwayFrom-uuid2"},
 						{Beacon: "weaviate://localhost/moveAwayFrom-uuid3"},
@@ -568,24 +568,24 @@ func Test_extractNearTextFn(t *testing.T) {
 					},
 				},
 			},
-			&searchparams.NearTextParams{
+			&nearText.NearTextParams{
 				Values:      []string{"transformed text", "c1", "c2", "c3", "transformed text"},
 				Certainty:   0.89,
 				Limit:       500,
 				Network:     false,
 				Autocorrect: true,
-				MoveTo: searchparams.ExploreMove{
+				MoveTo: nearText.ExploreMove{
 					Values: []string{"positive"},
 					Force:  0.5,
-					Objects: []searchparams.ObjectMove{
+					Objects: []nearText.ObjectMove{
 						{ID: "moveTo-uuid1", Beacon: "weaviate://localhost/moveTo-uuid2"},
 						{ID: "moveTo-uuid1", Beacon: "weaviate://localhost/moveTo-uuid2"},
 					},
 				},
-				MoveAwayFrom: searchparams.ExploreMove{
+				MoveAwayFrom: nearText.ExploreMove{
 					Values: []string{"epic"},
 					Force:  0.25,
-					Objects: []searchparams.ObjectMove{
+					Objects: []nearText.ObjectMove{
 						{ID: "moveAwayFrom-uuid1", Beacon: "weaviate://localhost/moveAwayFrom-uuid1"},
 						{ID: "moveAwayFrom-uuid2", Beacon: "weaviate://localhost/moveAwayFrom-uuid2"},
 						{Beacon: "weaviate://localhost/moveAwayFrom-uuid3"},
