+++ modules/text2vec-contextionary/neartext/text2vec_searcher.go
@@ -14,7 +14,7 @@ package neartext
 import (
 	"context"
 
-	"github.com/weaviate/weaviate/entities/searchparams"
+	"github.com/weaviate/weaviate/usecases/modulecomponents/nearText"
 
 	"github.com/go-openapi/strfmt"
 	"github.com/pkg/errors"
@@ -48,7 +48,7 @@ func (s *Searcher) vectorForNearTextParam(ctx context.Context, params interface{
 	findVectorFn modulecapabilities.FindVectorFn, cfg moduletools.ClassConfig,
 ) ([]float32, error) {
 	return s.vectorFromNearTextParam(ctx,
-		params.(*searchparams.NearTextParams),
+		params.(*nearText.NearTextParams),
 		className,
 		findVectorFn,
 		cfg.Tenant(),
@@ -56,7 +56,7 @@ func (s *Searcher) vectorForNearTextParam(ctx context.Context, params interface{
 }
 
 func (s *Searcher) vectorFromNearTextParam(ctx context.Context,
-	params *searchparams.NearTextParams, className string, findVectorFn modulecapabilities.FindVectorFn, tenant string,
+	params *nearText.NearTextParams, className string, findVectorFn modulecapabilities.FindVectorFn, tenant string,
 ) ([]float32, error) {
 	vector, err := s.vectorizer.Corpi(ctx, params.Values)
 	if err != nil {
@@ -95,7 +95,7 @@ func (s *Searcher) vectorFromNearTextParam(ctx context.Context,
 }
 
 func (s *Searcher) vectorFromValuesAndObjects(ctx context.Context,
-	values []string, objects []searchparams.ObjectMove,
+	values []string, objects []nearText.ObjectMove,
 	className string, findVectorFn modulecapabilities.FindVectorFn, tenant string,
 ) ([]float32, error) {
 	var objectVectors [][]float32
