+++ modules/text2vec-palm/neartext/searcher.go
@@ -14,6 +14,8 @@ package neartext
 import (
 	"context"
 
+	"github.com/weaviate/weaviate/usecases/modulecomponents/nearText"
+
 	"github.com/weaviate/weaviate/modules/text2vec-palm/config"
 
 	"github.com/go-openapi/strfmt"
@@ -50,11 +52,11 @@ func (s *Searcher) vectorForNearTextParam(ctx context.Context, params interface{
 	findVectorFn modulecapabilities.FindVectorFn,
 	cfg moduletools.ClassConfig,
 ) ([]float32, error) {
-	return s.vectorFromNearTextParam(ctx, params.(*NearTextParams), className, findVectorFn, cfg)
+	return s.vectorFromNearTextParam(ctx, params.(*nearText.NearTextParams), className, findVectorFn, cfg)
 }
 
 func (s *Searcher) vectorFromNearTextParam(ctx context.Context,
-	params *NearTextParams, className string, findVectorFn modulecapabilities.FindVectorFn,
+	params *nearText.NearTextParams, className string, findVectorFn modulecapabilities.FindVectorFn,
 	cfg moduletools.ClassConfig,
 ) ([]float32, error) {
 	// it is safe to call NewClassSettings even knowing that cfg can be nil, it
@@ -101,7 +103,7 @@ func (s *Searcher) vectorFromNearTextParam(ctx context.Context,
 }
 
 func (s *Searcher) vectorFromValuesAndObjects(ctx context.Context,
-	values []string, objects []ObjectMove,
+	values []string, objects []nearText.ObjectMove,
 	className string,
 	findVectorFn modulecapabilities.FindVectorFn,
 	settings localvectorizer.ClassSettings, tenant string,
